# openapi.yaml
components:
  schemas:
    Body_token_token_post:
      properties:
        client_id:
          default: ''
          title: Client Id
          type: string
        code:
          default: ''
          title: Code
          type: string
        grant_type:
          default: ''
          title: Grant Type
          type: string
        redirect_uri:
          default: ''
          title: Redirect Uri
          type: string
      title: Body_token_token_post
      type: object
    Jwks:
      description: A JSON Web Key Set as a dictionary.
      properties:
        keys:
          items:
            additionalProperties:
              type: string
            type: object
          title: Keys
          type: array
      required:
      - keys
      title: Jwks
      type: object
    LoginInfo:
      description: Data that is used to login as a user.
      properties:
        email:
          anyOf:
          - format: email
            type: string
          - type: 'null'
          description: e-mail address of the user
          title: Email
        name:
          description: the full name of the user
          title: Name
          type: string
        sub:
          anyOf:
          - type: string
          - type: 'null'
          description: subject identifier
          title: Sub
        valid_seconds:
          anyOf:
          - exclusiveMinimum: 0.0
            type: integer
          - exclusiveMinimum: 0.0
            type: number
          - type: 'null'
          description: seconds until the login expires
          title: Valid Seconds
      required:
      - name
      title: LoginInfo
      type: object
    OidcConfiguration:
      description: Data that is returned as OpenID Connect configuration.
      properties:
        authorization_endpoint:
          default: http://localhost:8080/authorize
          description: URL of the OP's Authorization Endpoint
          format: uri
          minLength: 1
          title: Authorization Endpoint
          type: string
        claims_supported:
          default:
          - sub
          - name
          - email
          description: List of the Claims that the OP can supply values for
          items:
            type: string
          title: Claims Supported
          type: array
        issuer:
          description: URL that the OP asserts as its Issuer Identifier
          title: Issuer
          type: string
        jwks_uri:
          default: http://localhost:8080/jwks
          description: URL of the OP's JSON Web Key Set document
          format: uri
          minLength: 1
          title: Jwks Uri
          type: string
        request_object_signing_alg_values_supported:
          default:
          - ES512
          description: List of JWS signing algorithms supported by the OP for Request
            Objects
          items:
            type: string
          title: Request Object Signing Alg Values Supported
          type: array
        scopes_supported:
          default:
          - openid
          - profile
          - email
          description: List of the OAuth 2.0 scope values that this server supports
          items:
            type: string
          title: Scopes Supported
          type: array
        service_documentation:
          default: https://github.com/ghga-de/test-oidc-provider
          description: URL of a page with information that developers might need to
            know when using the OP
          format: uri
          minLength: 1
          title: Service Documentation
          type: string
        token_endpoint:
          default: http://localhost:8080/authorize
          description: URL of the OP's Token Endpoint
          format: uri
          minLength: 1
          title: Token Endpoint
          type: string
        userinfo_endpoint:
          default: http://localhost:8080/userinfo
          description: URL of the OP's UserInfo Endpoint
          format: uri
          minLength: 1
          title: Userinfo Endpoint
          type: string
        userinfo_signing_alg_values_supported:
          default:
          - ES512
          description: List of JWS signing algorithms supported by the OP for the
            UserInfo endpoint
          items:
            type: string
          title: Userinfo Signing Alg Values Supported
          type: array
        version:
          default: 2.0.0
          description: Version of the test OpenID Connect Provider
          title: Version
          type: string
      title: OidcConfiguration
      type: object
    TokenResponse:
      description: The data that is returned by the token endpoint.
      properties:
        access_token:
          description: The issued access token
          title: Access Token
          type: string
        expires_in:
          description: Lifetime of the access token in seconds
          title: Expires In
          type: integer
        scope:
          description: The scopes granted
          title: Scope
          type: string
        token_type:
          default: Bearer
          description: Type of the token
          title: Token Type
          type: string
      required:
      - access_token
      - expires_in
      - scope
      title: TokenResponse
      type: object
    UserInfo:
      description: Data that is returned by the UserInfo endpoint.
      properties:
        email:
          description: e-mail address of the user
          format: email
          title: Email
          type: string
        name:
          description: the full name of the user
          title: Name
          type: string
        sub:
          description: subject identifier
          title: Sub
          type: string
      required:
      - sub
      - email
      - name
      title: UserInfo
      type: object
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.1.0
paths:
  /.well-known/openid-configuration:
    get:
      description: The OpenID discovery endpoint.
      operationId: get_openid_configuration__well_known_openid_configuration_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OidcConfiguration'
          description: Successful Response
      summary: Get the OpenID connect configuration
      tags:
      - TestOP
  /authorize:
    get:
      description: 'The authorization endpoint of the test OP.


        This authorizes the last logged in test user without checking credentials.'
      operationId: authorize_authorize_get
      parameters:
      - in: query
        name: response_type
        required: true
        schema:
          title: Response Type
          type: string
      - in: query
        name: client_id
        required: true
        schema:
          title: Client Id
          type: string
      - in: query
        name: redirect_uri
        required: true
        schema:
          title: Redirect Uri
          type: string
      - in: query
        name: scope
        required: true
        schema:
          title: Scope
          type: string
      - in: query
        name: state
        required: true
        schema:
          title: State
          type: string
      responses:
        '302':
          content:
            application/json:
              schema: {}
          description: Redirected back to client.
        '422':
          description: Validation error in submitted data.
      summary: OIDC Authorization Endpoint
      tags:
      - TestOP
  /health:
    get:
      description: Used to test if this service is alive
      operationId: health_health_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: health
      tags:
      - TestOP
  /jwks:
    get:
      description: Get the JSON Web Key Set of the test OP.
      operationId: get_jwks_jwks_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Jwks'
          description: Successful Response
      summary: Get the JSON Web Key Set of the OP
      tags:
      - TestOP
  /login:
    post:
      description: Endpoint for logging in to the OP as a test user.
      operationId: login_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInfo'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                title: Response 201 Login Login Post
                type: string
          description: Access token has been created.
        '422':
          description: Validation error in submitted data.
      summary: Log in as a test user
      tags:
      - TestOP
  /token:
    post:
      description: The token endpoint of the test OP.
      operationId: token_token_post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_token_token_post'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          description: Access token has been granted.
        '400':
          description: Error in submitted data.
        '422':
          description: Validation error in submitted data.
      summary: OIDC Token Endpoint
      tags:
      - TestOP
  /userinfo:
    get:
      description: The UserInfo endpoint of the test OP.
      operationId: get_userinfo_userinfo_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
          description: User info has been fetched.
        '403':
          description: Not authorized to get user info.
        '422':
          description: Validation error in submitted data.
      security:
      - HTTPBearer: []
      summary: Get user information
      tags:
      - TestOP